requirements.txt
certifi==2018.1.18
dj-database-url==0.5.0
Django==2.0.3
django-heroku==0.3.1
djangorestframework==3.7.7
gunicorn==19.7.1
pipenv==11.9.0
psycopg2==2.7.4
pytz==2018.3
virtualenv==15.1.0
virtualenv-clone==0.3.0
whitenoise==3.3.1
